Ignore intermediate tags generated by git-svn.
git-svn creates intermediate tags (or branches) with names like
"tag@revision-copy" when a tag is created by copy from a directory not
matching another branch or tag. This happens if the copy is from a
subdirectory of trunk, a subdirectory of another branch or from a
directory outside the trunk/branch/tag layout.

For example, when indexing branchA, if branchA is created by copying
a subdirectory of trunk at revision 42, it will create the intermediate
remote 'brankA@42'. Such remotes are useless as 1) they do not exist in
Subversion, 2) git-svn will add branchA@42 as the parent of the next
commit on branchA, and thus its history will be part in branchA.

An alternative would be to do nothing, and waits for them to be removed
by Tags.clean() at the end (which removes tags not existing in Svn).
However this won't work because "git-rev-parse --verify" parses
references like branchA@42 as an ordinal specification (branchA@{42})
and exits with a non-zero error code.