Addresses PR comments.
This restores the command switcher from a9003d0a to allow multiple
commands to be run from the tool. For example, 'author' (to generate the
initial author list from Subversion) and 'clean-git' (to clean the
repository after the conversion).

The command 'clean' was renamed to 'clean-git' to prevent any confusion
with the SVN url parameter: it is not the SVN repository which is being
cleaned.