Improves handling of Subversion branchs/tags with invalid names.
Git is less lenient than Subversion regarding branch/tag names (for
example, it does not allow spaces in a branch name).

Git-svn works around this by URL encoding the ref names (for example,
'my branch' becomes 'my%20branches' when converted to Git). As a
result, when comparing the branch and tag names with Subversion after
the conversion, branches like 'my%20branches' are deleted because they
are not found in Subversion.

Thus this restores any URL-encoded character in a Git ref name by its
original (decoded) value prior to comparing it with its corresponding
branch or tag in Subversion.

For more information on the characters allowed in Git refs, see:
http://www.kernel.org/pub/software/scm/git/docs/git-check-ref-format.html