Makes the 'clean-git' command idempotent.
Previously the command would delete 'branchy' tags issued by git-svn
after converting them to proper annotated tags. This was to prevent
those 'branchy' tags to be converted to local branches later on (in
Branches.createLocal). As a result, the tags would not be regenerated
when running the command a second time.

To make the command idempotent, the original 'branchy' tags from
git-svn are now kept, but filtered out of the list of refs prior
to the creation of the local branches.