Fixes branches and tags after conversion.
Branches and tags with invalid characters like spaces or tildas will
remain URL-encoded after the conversion. For example, "my branch" in
Subversion will be "my%20branch" in Git after the conversion.

Thus the names of the Git branches and tags needs to be cleaned up
after conversion, _but only after_ URL decoding those names and
comparing them to Subversion in Branches.removeObsolete() and
Tags.removeObsolete(). This is because once the branch or tag is
cleaned up, the original name can be reconstituted. (For example, a
space or a tilda will be both converted to a hyphen.)

Thus, this adds Branches.fixNames() and Tags.fixNames at the end of
the scripts.