SVNGIT-223: improve handling of emails in Subversion names
The script used to remove any email prefix in the Subversions usernames,
for both OnDemand and non-OnDemand repositories. As a result, for an
OnDemand repository, the script would use an incorrect username to
fetch the user profile from JIRA.

Instead the script now cleanup any email prefix only for non-OnDemand
repositories, to generate the appropriate placeholders for the fullname
and email.